requirements.txt
flask
sqlalchemy
manim
openai-whisper
torch
librosa
